add gitattributes and move types to their own file
